Move shared promotion and copy helpers to domains
