+++ src/domains/pricing/bulkPromotion.ts
@@ -1,4 +1,4 @@
-import type { CatalogItem } from "../../domains/catalog/catalogItem";
+import type { CatalogItem } from "../catalog/catalogItem";
 
 export type BulkPromotionMode = "percent" | "amount";
